Make objects explicit again (#16)
* Make objects explicit again

If there is no type defined, any type is valid. In case of objects,
defining properties and stuff is not enough. Those are used only if
the respective document is an object, but non-objects are still valid.
Added explicit type definitions to all objects.

* Add forgotten type definitions

Some object definitions were still lacking an explicit type definition.
Added type: object to those.